Sync Memento with gs.
Misc fixes plus:

---
Memento: Improve memory squeezing.

It seems that Memory squeezing can sometimes run into
problems with the child hanging after a fork. This seems
to be related to the child process SEGVing and the
signal handler not picking it up.

I've introduced a workaround so that the parent now only
waits a maximum of 30 seconds before killing the child and
continuing.

---
Memento memory squeezing tweaks.

Add Memento_bt() - a function to output the backtrace
to stderr.

Call this when we fork for a memory squeeze.

Forking only clones the active thread, so we can only use
the forking memory squeezer on single threaded tasks. Even
with single threaded tasks, you can't unlock mutexes in
the child that were taken in the parent. This means that
we need to disable mutexes in the main application if we
want to memory squeeze safely.

Also tweak the time we spend sleeping while waiting for a child
to die. This makes a huge difference to the runtime.

If we die during squeezing due to a SEGV, then don't bother
listing the blocks that we leaked.

---
Memento: Add Memento_sequence

---
Memento: Fix type of signal handler function.

Reapply from gs. Due to conflicts this just actually renames
some params. Do this to get the 2 versions back into sync.

---
Update copyright notice with new head office address.

Also update copyright dates.

Remove gs_cmdl.ps as we no longer use it, and remove its entry from
psfiles.htm.

Remove xfonts.htm as this feature (xfont support) is long, long gone.

---
Memento: Add Mutex to protect Memento operations.

Solves problems that Ray was seeing while debugging the multithreaded
integration of our modified LCMS2.

---
Fix typo in Memento_breakpointLocked.

---
Memento: Fix typo in Mutex init.

Thanks to Shelly for pointing this out.

---
Fix Memento locking problem.

When calling Memento_event, it would lock, and then call
Memento_checkAllMemory which would lock again. This is fine
on windows, as locks are recursive, but not on Linux.