quick hack: fix infinite loop which will trigger for some PDFs and the 'mutool clean' command:
    'MUTOOL clean -gggg -D -c -s -AA "Sample-PDFs-for-format-testing/PDFIO-format-corpus/govdocs1-error-pdfs/error_set_2/219789.pdf" "__mujstest/Sample-PDFs-for-format-testing/PDFIO-format-corpus/govdocs1-error-pdfs/error_set_2/%04d-219789.clean.pdf"'

-- causes infinite loop in the code: hits the 'else if (rune == 32)' branch inside 'static void walk_string(fz_context *ctx, int uni, int remove, editable_str *str)' and loops indefinitely at the same spot as str->pos is not updated.

Quick fix which I applied is to 'skip' that rune space, i.e. keep it around.

WARNING: after testing the initial fix, I added code to check for end-of-string (rune == 0) as that MUST NOT result in an illegal continue due to uni == 0.

The walk_string() code was done in commit SHA-1: abcb3e68670ebc2e5127953462a026fe1a5dd321 :: Improve redaction capabilities.

When redacting text from a PDF file, we want to ensure that we
redact the same text in the StructTree parts.

Add 'structparents' param to the text filter creation; this should
be the number given in the content stream's "StructParents" entry
(or -1 if there isn't one).

Within the PDF filter processor itself, we move to keeping track
of current tags as well as pending ones. As we encounter each tag,
we look for the /MCID object, if there is one, and we store a
reference to this in the tag_record. From this, we look for /Alt
and /ActualText tags, and, if present, extract the strings from
them as UTF8.

Whilst filtering, whenever we encounter a char, we consider the
/MCID's that are in force; if we have Alt or ActualText strings
we check whether the next character agrees with the character
pulled from the file or not. If the strings agree, and we have
to remove a character, we remove it from the Alt/ActualText strings
too. If the strings disagree (and characters are removed), then
we err on the side of caution and remove the entire string.

When we come to end the marked content (with an EMC), we update
the edited strings in the file.