Now all tools and mupdf code uses fz_error(), fz_warn() and fz_info() instead of fprintf(stderr, ...) --> we have achieved our goal of a silent stderr in mujstest, while all error/warn/info output is fed into a logfile in our test case.
NOTE 1: partial lines, which are not supported by fz_info() et al, are tagged with a "\v" at the end.

NOTE 2: many error lines have been LF-stripped while the new fz_error() et al code checks for a trailing LF in the message and strips it off as the output print code will append a LF of its own, for any channel (such as stdout or stderr) which needs it. This allows us to pipe error/warn/info output to channels which DO NOT work well will terminating LFs to separate lines, e.g. GUIs.