Coverity 215307/215308: Explicitly limit ranges of read fields by masking.
It appears that Coverity is complaining that we do not validate the ranges
of the JBIG2 header fields we read. When one read and byte swapps four
bytes into a 32 bit entity Coverity really ought to be able to deduce the
range of that entity, but it doesn't. Even if it is silly to mask the 32
bits we just read with 0xFFFFFFFF let's just do it. Doing it this way beats
any kind of Coverity specific comments.