Bug 702898: Fix SVG output clipping issues.
When we define a pattern, we do so using the SVG symbol mechanism.
Some SVG renderers, by default, make a clipbox around the rendered
symbol. This is no good for contents of patterns that are offset
and should still appear correctly due to the repeats in the patterns.

Accordingly, we make every such symbol have the style="overflow:visible"
attribute.