augment pdfinfo tool: output additional metadata info.
also make the output slightly easier to unambiguously parse by machine (even when you feed it PDFs with 'evil' titles containing, say, a '>>' character sequence: those strings are now encoded the same way as was already done in the fmt_xxx() internal debugging APIs.

augmented the debug/diagnostic pdf_print_obj API (and all its relatives): when a *indirect* node is encountered, this is printed as before. However, now it also executes a `RESOLVE()`, just like all the other pdf_object APIs already do under the hood. The important bit being that here the 'resolve' is displayed step by step, including the sanity check and error line (when a cycle is detected/assumed), plus a dump of the indirectly referenced object itself.
This code is used by the pdfinfo augmentation, where the complete Info meta dictionary is dumped as part of the info output: when applications have added their own metadata keys to that dictionary, we will be able to see that!