- fix a couple of crashes uncovered by bulk testing `mutool pdfinfo`
- fix/revert commit SHA-1: bdd7c9764f0064395b3c8058746eb2931c245c2a :: augmented the debug/diagnostic pdf_print_obj API (and all its relatives)

That code is a generic code base to output Postscript which we just *nuked* by adding that.

The fix is two-fold: the metadata dict dump (where this all started) is now chain-resolved *before* going to the print routines. (Turns out those indirect references weren't all that interesting after all.)
Second, we patch the string writer: when in ascii mode, writing non-ascii characters (such as copyright characters, seen many times in our test set), the output would be illegible hex dumps.
The patch keeps the Postscript hexdump, but *adds a Postscript comment* at the end of that line, which then has the "semi-ASCII" string printed in a way that's slightly more legible to humans. TODO: eff the octal encodes in there, for it's human perusal only and the subset of humans who can read octal fluently is... dwindling.