complete fix for the run-away outline loading for https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4835977/ and other PDFs which have arbitrary deep outline levels of single nodes with same-named single children.
Notes:

we check both parent AND grandparent for title, page and URI identity before we decide to 'ignore' the current single node, as it would only add another *useless* outline level.
We check both parent AND grandparent to prevent any possible match with *sane* PDF outline structures: some *legal* PDFs have outline structures where the parent and *first child* have the same title, page and (internal) uri -- these PDFs usually have multiple children at that second level, but we are better safe then sorry while we 'ignore' outline nodes.

Added benefit of this fix/tweak is that both the call stack and exception stack depths are significantly reduced. Nevertheless, we decided to keep the (exception) stack adjustments made in the previous commit.

The result is a reasonably "sane" outline, while some of the "insanity" remains visible, but that's okay in my book: the user/reader can still observe a hint of trouble in the original PDF that way, without run-away outlines / viewers (Foxit Reader shows it like it is which makes for a very weird outline view side panel; Chrome also doesn't seem to like this sort of thing as the CPU was cooking for a while before rendering the PDF pages...)

N.B.: of course one can still game the system with specially crafted attack PDFs, e.g. one where the titles for parent and child are alternated to thwart `are_outline_nodes_equal()` but this at least cleans up after crufty PDFs which have been found in the wild.

# Conflicts:
#	source/pdf/pdf-outline.c