Bug 703654: Mark colorspace variable with fz_var.
OSS-fuzz discovered a memory leak of the cs variable in load_devicen().

The reason is that the cs variable is not marked with fz_var, it is set
within fz_try(), yet it is used in fz_catch() to drop the colorspace upon
exception. Since it is not marked as fz_var the will be reset to NULL
when the exception happens inside fz_try(). This means that dropping cs in
fz_catch just drops NULL, thus leaking the allocated colorspace.

Simply marking the variable with fz_var fixes the issue so that the
cs variable is not reset to NULL upon exception.