fix O(n^2) costly behaviour for corrupted PDFs which have a damaged page list, e.g. `W:/Sopkonijn/!QIQQA-pdf-buffer-dir/__possibly_erroneous/EBA9C2BAC1E301AEFA99296E0BF43CEBDA4D175.pdf`
The key here is that when you loop through the pages, reporting or rendering each one (as is done by the `mutool metadump` utility, for example, then once you hit the corrupted page entry (413 in this example), you'll get error reports about the object being out of bounds. Fine, but with the original code, several pdf_dict_get_inheritable() calls would be executed for this b0rked 'kid' slot, resulting in *multiple* error reports plus exceptions handled for a *single* corrupted slot.

Now, in the given PDF example, the corruption is not a single slot, but everything from slot 413 onwards. Let's call that N.

Also note that the classic code did return the b0rked kid slot as a HIT reference, resulting in yet more out-of-bounds reports and exceptions handled while the process attempts to access page aspects for further processing/reporting, making one b0rked slot already pretty costly.

Then there's the outer-most loop, which was going through the pages, attempting to report on every page. From 413 onwards, in the example, that's bound to fail as those page slots are all b0rked, resulting in a total b0rked page slot access cost of O(0.5 * N * N) -- 0.5 because the number of b0rked slots hit increases as the page number to look up increases. (Keep in mind that O(1) means many error reports as those occur for a single page slot test, resulting in a high base cost that's multiplied by that N^2 square!)

Anyhow, for the given example PDF and `mutool metadump` as a processing tool, this resulted in a run-away cost of many *minutes*!

The new code adds a tiny up-front cost cost to the page list testing loop by performing the checks that detect a b0rked page slot *before* the pdf_dict_get_inheritable() et al APIs get to it and start b0rking errors and exceptions instead. When this new 'sanity checking' code hits such a snag, it does something extra:

- it scans the full remainder of the page table to detect all b0rked slots at once, so the error reports for each page slot in there are nice, sequential and tight together.
- it also *drops* the b0rked slots and replaces them with PDF_NULLs, which serve as fast-track b0rk detect & handling when the page table search continues and/or the page lookup API is revisited, e.g. when the application code attempts to access a next page.
- instead of producing a reference to a known-b0rked page slot `pdf_obj*`, the code returns a NULL hit reference: as the page we were looking for is unreachable anyway, this removes many later error reports due to code attempting to dereference that b0rked page slot reference and results in cleaner execution overall.

Added maintenance cost of this solution is the notion that the "sanity checking" code section is an optimized, reduced duplication of the code later in the page list traversal loop, while the optimized sanity check code has precise knowledge of the checks and error reports produced and other behaviour by the pdf_dict_get_inheritable() API calls which would otherwise have been invoked. The optimized code chooses explicitly to behave *different* by ultimately producing a NULL instead of a `hit` reference, as is stated in the code comments too, but this does not take away the fact that this is some stuff that's been unrolled from those APIs. Any change to these APIs will therefor possibly need to be replicated in the "sanity checking" code. This is deemed safe enough as those APIs and internal behaviour have been stable for many years and there's no observable need to change the internal behaviour of those pdf_dict_get_inheritable() et al APIs any time.

The *benefit* of this change is very fast recovery and rapid performance for b0rked PDFs such as the example: `mutool metadump` produces the same amount of info in a couple of seconds instead of taking 10+ minutes and producing a huge error logfile to boot: the "full page list scan" in the optimized code has specifically been added to remove those many thousands of error report duplications.