setup.py: allow generation of .whl on systems without clang-python and/or swig.
When generating a sdist, we run './scripts/mupdfwrap.py -b 0 2' to build C++ source and SWIG source, and include this source in the sdist.

When generating a wheel, if $MUPDF_SETUP_HAVE_CLANG_PYTHON is set but not '1', we don't generate C++ code, instead we expect it to be already in place.

Similarly if MUPDF_SETUP_HAVE_SWIG is set but not '1', we don't run SWIG to generate C++ code, instead expecting it to already exist.