Use FZ_DATA_DECLARATION and FZ_DATA_DEFINITION.
On Windows, users of DLL must build with FZ_SHARED defined.

We don't actually use FZ_DATA_DEFINITION anywhere - it looks like we need
to declare global data with __declspec(dllimport) when building client
code. For building the DLL itself, data does not need to be qualified either in
declarations or definitions. (So FZ_DATA_DEFINITION is never actually used.)

We still get a warning for mupdf::metadata_keys.