scripts/pypackage.py: generic support for building/uploading Python packages.
Allows easy building of sdist and wheels on local and remote linux/windows
machines, plus uploading of sdists and wheels to pypi, and testing that 'pip
install' works.

On Linux we build wheels inside a manylinux docker container. On Windows we use
the 'py' command to select which installed Python to use.

We use venv when running Python tools such as pip so we can upgrade, install
twine, and choose which python to use.

We use jlib.Arg for command-line parsing.