scripts/pypackage.py: generic support for building/uploading sdists and wheels.
Allows easy building of sdist and wheels on local and remote linux/windows
machines, plus uploading of sdists and wheels to pypi.

On Linux we build wheels inside a manylinux docker container. On Windows we use
'py' to select which installed Python to use.

We use venv when running Python tools such as pip so we can upgrade, install
twine, and choose which python to use.