scripts/jlib.py: allow handling of login/password prompts etc with jlib.system().
If caller needs prompts that don't end with newline to be displayed, pass
bufsize=0 to jlib.system().

We then pass bufsize=0 to subprocess.Popen(). But we also need to do our
own decoding because subprocess's universal_newline and codec.streamreader
both seem to do their own internal buffering. What works is to use a
codecs.getincrementaldecoder and pass it the raw bytes that we read from the
child process.

Also added <raw> arg to jlib.log() and jlib.log_string() to show output from
jlib.system() that is not line-based; also added _log_text_line_start so we
keep track of whether we are in middle of unterminated line when in raw mode.