Bug 704045: Only flag annotation as changed when setting appearance.
Previously when setting an appearance using a contents buffer the
annotation was flagged for resynthesis, but resynthesising would
mean incorrectly replacing the appearance that was just set.

Moreover, if the appearance was set using a display list, not only
was the annotation flagged for resynthesis, but the document would
also be flagged as dirty. There should be no difference between
setting the appearance using a contents buffer or a display list
apart from how the appearance stream contents are provided.

Finally, when the appearance for an annotation is set it must be
flagged as having changed so that pdf_update_annot() will inform
the app that the annotation needs to be redrawn.