Bug 704045: Do not consider missing entries in local xref as local objects.
Consider a document that has a number of objects, next a local
xref is created and some local objects are added. With the prior
source code any of the originally existing objects (none of which
are present in the local xref!) would be considered local as well.

Starting from this commit MuPDF now considers non-present objects
or free objects not to be local.