mupdf-gl: Ensure update_title called correctly on document dirty/clean.
The existing code watches for the value pdf_has_unsaved_changes()
changing around the main edit loop, and calls update_title when
a change is spotted. Unfortunately this doesn't capture edits made
using dialog boxes (such as tx_dialog and ch_dialog).

Therefore move to a simpler version where we record the state in a
static.