prepwork for hOCR/HTML output with separate image files, using URIs in the generated HTML.
After a few experiments, it looks like the cleanest (and least impact on the existing API) approach is to enhance the `fz_output` structure and API: the fz_output will need to remember/keep the filepath of the primary output file, which will be used as a template (or part thereof) for the secondary files, which would be the extracted images for a PDF-to-HTML dump.

As the 'secondary' outputs will be zero-to-many (0 or more images), the `fz_secondary_outputs` structure and APIs are designed to help create filenames and extra `fz_output` instances, one at a time, while also offering help to ensure we're creating *unique* (non-colliding) image paths, which can then be written as image URIs into the primary input.

That also means the image filename/path must be made unique and *used* / *claimed* at the time the image URI is written into the primary output! This is why we cannot get away with a two-rounds-through-the-PDF kind of approach, where the first round would generate the content (or the images) and the next round would export the other parts: any outside process can create any of the file paths and we'ld have a collision. -- Though *first* creating all the image files, memorizing their paths for use with the second round, would solve this problem. But then the second round would be pretty similar in complexity and API changes, compared to what we currently are setting up with the `fz_secondary_outputs` APIs.

The design goal is to make the `fz_secondary_outputs` APIs flexible enough that those can handle both SMIME/HTML style outputs and HTML/IMG/CSS multifile outputs: the former can be accomplished by buffering the images and other sub-filees in a secondary stream, which is then appended at the end when the main HTML content filee has been completely written. Do note that this is currently only an *idea* and has not been implemented yet.