Bug 704247: Ensure error jmp_buf's are aligned.
We align the error jmp_buf's to FZ_JMPBUF_ALIGN bytes. By default
this is 32, but can be increased by predefining FZ_JMPBUF_ALIGN
to a larger number. This must be a power of 2.

I hope this will solve the bug, but I can't be sure as I do not have
access to a Solaris system to test on.