scripts/mupdfwrap.py: allow choice of python or c# bindings.
Fixed python support broken by C# experiments. Now we only include the swig .i
C++ helper code for handling out-params if we are building for python.

If we build both python and c# bindings into _mupdf.so, python test works, but
c# test does not - it looks like C# fails to load the .so.

C# works ok if we build just c# bindings into _mupdf.so.