scripts/mupdfwrap.py: wip out-params for windows.
get_args(): added extra detection of non-out-param FILE* because
get_base_typename() on windows appears to convert FILE to __sFILE (maybe
because FILE is a typedef?).

make_outparam_helper_csharp(): allow use where C#-friendly names such as
'unsigned long' are different from 'size_t'. This happens because we generate
C++ on Unix for use on Windows. Instead of passing the addresses of members of
the *_outparams struct, we pass addresses of local variables such as 'size_t
foo' and copy them into the struct before returning.

Things still fail on Windows because v4.0.30319/csc.exe appears not to be able
to handle fns that return a tuple - max langversion is 5?