scripts/mupdfwrap.py: improved generated code a little.
Only generate C# out-param helpers for fns that will be c# class methods -
these need to be in a top-level C# class in order for the extension method to
work, and so for non-class methods we'd want to handle things differently, e.g.
putting things in a top-level mupdf namespace instead, which seems more trouble
than it is worth.