scripts/mupdfwrap.py: fixed bug in generation of fz_link iterators.
We were creating a LinkIterator by passing a fz_link*, which was assumed
to be an owned pointer when passed to LinkIterator's 'Link m_item' member,
so we weren't incrementing the reference count.

The fix is to for Link::begin() to pass *this to LinkIterator's constructor
instead of m_internal.