WIP Return mark for each new search hit.
With the marker each individual hit on the page can be identified.
A quad that starts a new array has a matching 1 in the 'hits' array.
Each continuation quad is marked with a 0 instead.

This is the minimally invasive API change, it just adds an optional
extra int array for recording the hit markers.

The Java and mutool run APIs have been changed to return an
array of array of quads for the search results.