further work on https://bugs.ghostscript.com/show_bug.cgi?id=701945 : now userland code can configure the maximum nodes and/or maximum time (msecs) to be spent on a page render via the `cookie` object.
The limits currently only apply to the list-based default render (as available via the `mudraw` utility) and, once triggered on a page, will cut that page render short, paint a notice over the page to notify the user/viewer, and continue rendering the next page, if desired.

The previous limiter code has been clened up and the tracking parameter in the `fz_context` object has been removed as part of the cleanup.