Bug 704704: Fix mutool banding issues.
A while ago, I tried a commit to solve this. I now look
back on that code, and can't see that it makes any sense.
So, revert commit eaec63901caa0e72009528f430b790e8b87aac1e
here, and try a new fix.

The problem was not in the draw_device implementation, but
rather in how we were calling it. Currently we feed in the
same bitmap stripe each time (y=0 to y=50 say), and alter
the ctm each time. This works fine for everything other than
patterns, whose phase is determined by the position on the
page.

Accordingly, we update mutool to pass in the bitmap stripe
with different spatial coords each time, and leave the ctm
alone. This correctly allows the draw device to position the
patterns.

# Conflicts:
#	source/tools/muraster.c