Bug 704802: Fix problem with reloading journal.
The bug describes a scenario where the journal fails to reload.

The initial file in the bug contains a single XRef with 1036 objects.
We load it, and edit the "Employee Contents" field to be "test". We
then save it incrementally. This saves us a file with an incremental
update. It's a new style incremental update that comes in 3 sections;
[(start=1036, len=2), (start=553, len=1), (start=265, len=1)].

Next, we restart mupdf-gl and reload that file. This loads the file
into memory, and reloads the xrefs. It loads these by running backwards
through the xrefs, so starts with the 3 part one first.

It loads this as is (keeping it as a 3 part linked list). Then it loads
the preceeding section (the 1036 object one). Unfortunately, because
this is a "new style" xref, it looks for the 'object entry' for the
new style xref object as part of opening the stream. This is an object
we are still loading, so doesn't actually have an entry yet.

Accordingly, our code decides that it must be a new object to be put
into the xref, and so 'solidifies' our 3 section xref into a single
larger section.

Continuing our test, we edit the "Supervisor Comments" to say "test2".
This creates a new incremental section, and changes some objects there.
Then we snapshot the file.

Thus, we come to 'fingerprint' the file to save a journal. The
fingerprint is based upon the 2 xref sections that aren't the current
incremental one. Thus, one of 0-1036, and one of 0-1038.

We then exit mupdf-gl, and reload the snapshotted file. This time,
as we load the xrefs in, we first read the new 'incremental' xref, then
the 3 part xref, then the 1 part xref. When the 'solidification' is
triggered as we load the 1 part xref, it instead solidifies the
incremental xref.

Next the code loads the journal. The file is refingerprinted, and it
fails to match (due to the xref sections appearing to be differently
laid out), and the journal load fails - hence undo is disabled.

The fix is to tweak the functions used so that we don't inadvertently
trigger the solidification as we load the xref. This is the change
to pdf-stream.c.

Fixing this uncovers another problem, whereby the incremental xref
loaded from the snapshotted file wasn't appearing as a solid one.
(Incremental xrefs held in memory are assumed to be solid). Accordingly,
we have to add a new call to pdf_ensure_solid_xref in
pdf_deserialise_journal to ensure this is the case.

This in turn uncovers another problem, whereby pdf_ensure_solid_xref
was solidifying the oldest xref, rather than the newest. Fix that
here.