Bug 704930: Check for Appearance Stream reuse.
We need to ensure that an object for an appearance stream, is not used
from more than one annotation before we reuse it.

Paul came up with the nice observation that any new appearance stream
we generate will go into the incremental section, and we'll never
put shared ones into that section. So, if and only if the appearance
stream from an annotation is in the incremental section, we can reuse
it.

# Conflicts:
#	source/pdf/pdf-xref.c