jni: Expose an activation API for known widgets.
Previously PDFPage.activateWidgetAt() had an API to trigger the
sequence of events for the first widget at a a given coordinate.

If the app already has determined what PDFWidget to operate
on, looping over the list of widgets is unnecessary.

Since Acrobat also allows annotations to process events this
change is done for PDFAnnotation which PDFWidget extends.