Allow updating a field value when ignore_trigger_events==1
This pertains to https://bugs.ghostscript.com/show_bug.cgi?id=705160
The field in question has a value (3.1415) that's different than
its formatted value (3.14).

When we edit a field, we first set ignore_trigger_events=1
so as we type, mupdf renders the field without formatting. This works
fine while we type. But until we type, the field still looks formatted.

I want to set the field to its current value once, to get mupdf to render
it unformatted, before we begin typing. But update_field_value
is checking to see if the value it was given is the current value.

In this change we only do that check if (!ignore_trigger_events).