gl: Invalidate ui when console is updated.
Previously when not invalidated, if the console contents were updated
by e.g. a PDF event callback the new contents would not show up until
the mouse pointer was moved.

Adding the ui_invalidate() calls to gl_js_console_*() requires that
the ui_invalidate() will not try to update the glut window until it
has been created. The reason is that the document and page are loaded
(hence event callbacks for formatting can be executed) before the
glut window is created.