Introduce fz_html_story and associated API.
Rejig html layout/draw code. Comment it. Fix some issues with
vertical margin collapse. Add restartability to both layout and drawing.

Example file to show margin collapse problems: (Replace &hash;
with the obvious :) )

<!DOCTYPE html>
&hash;a { margin: 30px; }
&hash;b { margin: 20px; }
&hash;c { margin: 5px; }
&hash;a { border: 1px solid red; }
&hash;b { border: 1px solid green; }
&hash;c { border: 1px solid blue; }
<style>
</style>
<body>
<div id="a">A</div>
<div id="b"><div id="c">C</div></div>
</body>
</html>

Introduce API to create a story, to place it (several times possibly),
to draw it, and to drop it.