HTML restarting layout: swallow top margin on restart.
When laying out a second or subsequent rectangle from a text story
swallow any margin at the top. This means that new paragraphs
starting the second block will start at the same vertical position
as lines broken from the middle of a paragraph.

I'm not 100% sure this is correct, but in discussion with Tor, we
both felt it was better. I've kept this as a separate commit so
we can change our minds easily later.