HTML layout: Add "leading" css property.
Add leading to the properties in css-properties.gperf and rebuild
css-properties.h from that.

Update the CSS property routines to cope with 'undefined' properties.
The leading property is initially undefined, and only set when met
within the file.

Update the layout routines to honour leading. If leading is set, then
we always 'advance' lines by 'leading', rather than the maximum of the
lineheight, and the text height. We still position the text according
to the lineheight.