Honour ExtGState TK flag.
Collecting the TK (Text Knockout) flag in the PDF interpreter is
easy. Dealing with it is somewhat harder.

The idea is that if the TK flag is set, every text object (BT/ET)
should enclose its contents in a knockout group. This both protects
separate strings in the object (e.g. (foo) TJ (bar) TJ) from
nasty effects if they overlap each other, but also individual letters
from overlapping other letters in the same string.

Accordingly, this means we need to push/pop a knockout worker group
around every single letter in each string in the worst case.

An idea that initally seems nice is to not bother pushing the knockout
group if we don't have any alpha values in play - unfortunately the
alpha values (and text rendering mode etc) can change during a BT/ET,
and hence pushing a knockout group when the first non-solid alpha
level is seen will not work.

Accordingly, the best we can do is to not push the knockout group if
there is no alpha at all on the page.