Bug 705677: Fix rendering of zero length dash line caps.
This bug stems from what appears to be a false memory I had of
the way zero length dash segments should be capped.

The 1.7 spec says:

> If a subpath is degenerate (consists of a single-point closed path
> or of two or more points at the same coordinates), the S operator
> paints it only if round line caps have been specified, producing
> a filled circle centered at the single point. If butt or projecting
> square line caps have been specified, S produces no output, because
> the orientation of the caps would be indeterminate. (This rule
> applies only to zero-length subpaths of the path being stroked,
> and not to zero-length dashes in a dash pattern. In the latter
> case, the line caps are always painted, since their orientation
> is determined by the direction of the underlying path.)

A rephrased version appears in the 2.0 spec:

> If a subpath is degenerate (consists of a single-point closed
> path or of two or more points at the same coordinates), the S
> operator shall paint it only if round line caps have been
> specified, producing a filled circle centred at the single
> point. If butt or projecting square line caps have been
> specified, S shall produce no output, because the orientation
> of the caps would be indeterminate. This rule shall apply
> only to zero-length subpaths of the path being stroked, and
> not to zero-length dashes in a dash pattern of a non-degenerate
> subpath. In the latter case, the line caps shall always be
> painted, since their orientation is determined by the
> direction of the underlying path except in the case of a
> degenerate subpath. A single-point open subpath (specified
> by a trailing m operator) shall produce no output.

My interpretation of this is that zero length dash segments that
have 1) round caps should be displayed as dots, 2) Butt caps
resolve to a zero sized area (so even if they are 'drawn' they
won't actually mark the page in AA modes), and 3) Square caps
(projected caps) should be displayed as squares aligned with the
direction of the path.

Previously, I'd been under the mistaken impression that all such
line caps should be displayed as being round.

Here we modify the code to fix the issue.

In order to satisfy the final sentence quoted from the 1.7 spec,
we are required to keep the 'direction' of the path separate
from the displacement for each segment.