further work on https://github.com/sumatrapdfreader/sumatrapdf/issues/1030 / https://bugs.ghostscript.com/show_bug.cgi?id=701945 Turns out that mupdf goes nuts creating (and then blending+dropping) knockout layers for every path command and this particular PDF is chockful of *paths*, so it takes *ages* to render (tested again with latest codebase and modern AMD Ryzen machine: we did not wait for the original result to be produced (> 5 minutes runtime).
With this 'fix' (which kills the knockout layer creation and subsequent/consequent blend+drop) the render time is multiple seconds on this hardware, i.e. tolerable/acceptable. The trick is to keep all knockout layer construction EXCEPT for anything that's directly PATH rendering involved: I assumed that would be too pedantic/precise to have and the inspected render result compares reasonably with FoxIt output, for example -- only did an eyeball check, not a mechanic image diff, as I was hunting for a *reasonable* speed-up of the code without having to dig deep and do a wild refactor.